+++ src/catch2/meson.build
@@ -144,7 +144,6 @@ internal_headers = [
   'internal/catch_test_failure_exception.hpp',
   'internal/catch_test_macro_impl.hpp',
   'internal/catch_test_registry.hpp',
-  'internal/catch_test_run_info.hpp',
   'internal/catch_test_spec_parser.hpp',
   'internal/catch_textflow.hpp',
   'internal/catch_to_string.hpp',
@@ -185,6 +184,7 @@ internal_headers = [
   'catch_template_test_macros.hpp',
   'catch_test_case_info.hpp',
   'catch_test_macros.hpp',
+  'catch_test_run_info.hpp',
   'catch_test_spec.hpp',
   'catch_timer.hpp',
   'catch_tostring.hpp',
